services/data_generation/requirements.txt
annotated-types==0.7.0
confluent-kafka==2.12.2
numpy==2.2.6
pandas==2.3.3
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
